requirements.txt
# Core Deep Learning and Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# API and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Configuration and Data Handling
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0

# File handling
pathlib>=1.0.1
